Hack around other Finalizers adding POD sections
If other Finalizers ran after this plugin, and those Finalizers added
sections, then this plugin would silently fail to detect if those
added sections were duplicates. Of course,
Pod::Weaver::Section::Leftovers, the plugin that is by far most likely
in practice to produce duplicate sections, happens to be a Finalizer.
To hack around this, we now do the Preparer role so that we can go in
before any Finalizers are run and move ourselves to the end of the
weaver object's plugin list.

I apologize in advance.

Thanks to CJM for providing simple and non-broken plugin-reordering
code.